+++ custom_components/intex_pool/const.py
@@ -46,6 +46,7 @@ PLATFORMS = [
     Platform.NUMBER,
     Platform.BUTTON,
     Platform.SELECT,
+    Platform.TIME,
 ]
 
 # --- device types ---